+++ ICD.Connect.Cameras.Mock/Properties/AssemblyInfo.cs
@@ -4,4 +4,4 @@ using System.Reflection;
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Cameras.Mock")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2020")]
-[assembly: AssemblyVersion("4.3.0.0")]
+[assembly: AssemblyVersion("4.4.0.0")]